aider: Increase the font size of the label on the image in image_rec.py.
# Aider chat conversation:

USER: make a bigger font on the label

ASSISTANT: To assist you better, I need more information. Could you please specify the programming language and the framework you are using? Also, could you provide the file that contains the label whose font size you want to increase?

USER: I added these *read-write* files: image_rec.py

ASSISTANT: To increase the font size of the labels drawn on the image, we need to modify the `image_rec.py` file. Specifically, we need to change the `ImageFont.load_default()` method which loads the default font. This method doesn't allow us to specify the font size. Instead, we can use the `ImageFont.truetype()` method which allows us to specify the font and its size.

However, to use `ImageFont.truetype()`, we need a TrueType or OpenType font file. Python doesn't come with a default TrueType font, so you'll need to provide one. A commonly used font is Arial, which is typically available on many systems.

Here's how you can modify the code:

```python
image_rec.py
<<<<<<< ORIGINAL
    font = ImageFont.load_default()
=======
    font = ImageFont.truetype("arial.ttf", 15)
>>>>>>> UPDATED
```

In the updated code, `"arial.ttf"` is the name of the TrueType font file and `15` is the font size. You should replace `"arial.ttf"` with the path to the TrueType font file you want to use, and `15` with the desired font size.

Please note that the path to the font file should be either an absolute path or a path relative to the location where the Python script is run. If the font file is in the same directory as the Python script, you can just use the font file name as the path.

If you don't have a TrueType font file, you can download one from the internet. There are many websites that offer free TrueType fonts.